# generate_early performance improvements