--- a/worlds/jigsaw/__init__.py
+++ b/worlds/jigsaw/__init__.py
@@ -138,6 +138,7 @@ class JigsawWorld(World):
         
         merges = 0
         clusters = []
+        added_piece_count = 0
         
         self.precollected_pieces = []
         self.itempool_pieces = []
@@ -158,7 +159,7 @@ class JigsawWorld(World):
                     elif self.options.piece_order == PieceOrder.option_every_piece_fits:
                         for i in range(len(pieces)):
                             p = pieces[i]
-                            c, m = add_piece(clusters, p, self.nx, self.ny)
+                            c, m = add_piece(clusters, p, self.nx, self.ny, added_piece_count)
                             if first_piece or m > merges:
                                 pieces.remove(p)
                                 break
@@ -170,7 +171,7 @@ class JigsawWorld(World):
                         best_result = 5
                         for i in range(len(pieces)):
                             p = pieces[i]
-                            c, m = add_piece(clusters, p, self.nx, self.ny)
+                            c, m = add_piece(clusters, p, self.nx, self.ny, added_piece_count)
                             if first_piece or m - merges <= best_result_ever:
                                 best_piece = p
                                 best_result = 0
@@ -192,7 +193,8 @@ class JigsawWorld(World):
                 else:
                     self.precollected_pieces.append(p)  # if no merges left, add piece to start_inventory
                     
-                clusters, merges = add_piece(clusters, p, self.nx, self.ny)  # update number of merges left
+                clusters, merges = add_piece(clusters, p, self.nx, self.ny, added_piece_count)  # update number of merges left
+                added_piece_count += 1
                 
                 first_piece = False
                     
@@ -200,18 +202,21 @@ class JigsawWorld(World):
         self.actual_possible_merges = [0]
         merges = 0
         clusters = []
+        added_piece_count = 0
         
         for p in self.precollected_pieces:
-            clusters, merges = add_piece(clusters, p, self.nx, self.ny)
+            clusters, merges = add_piece(clusters, p, self.nx, self.ny, added_piece_count)
             self.possible_merges.append(merges - self.options.number_of_checks_out_of_logic.value) 
             self.actual_possible_merges.append(merges)
+            added_piece_count += 1
         for c, p in enumerate(self.itempool_pieces):
-            clusters, merges = add_piece(clusters, p, self.nx, self.ny)
+            clusters, merges = add_piece(clusters, p, self.nx, self.ny, added_piece_count)
             if len(self.itempool_pieces) - c < 10:
                 self.possible_merges.append(merges)   
             else:
                 self.possible_merges.append(merges - self.options.number_of_checks_out_of_logic.value)   
             self.actual_possible_merges.append(merges)
+            added_piece_count += 1
         
         self.pieces_needed_per_merge = [0]
         for i in range(1, self.npieces):