## What is this fixing or adding?

Improves the performance of `generate_early()`, particularly for least merges possible order and to a lesser extent every piece fits order. Random order is close to unaffected by these changes.

### [Keep clusters as sets instead of converting to lists and back again](https://github.com/spineraks-org/ArchipelagoJigsaw/commit/ad160a01761df8529cb58cd95e4ae95b33aeb952)

This takes a 1000 piece 100% strictness least merges order puzzle's generate_early from 15s to 10s for me, and produces identical output with the same seed.

### [Use set.isdisjoint instead of creating intersection sets](https://github.com/spineraks-org/ArchipelagoJigsaw/commit/11674e599825ce782e894ef81321a700a853c132)

When only the knowledge of whether two sets intersect is needed and not what the actual intersection is, negating the result of set.isdisjoint can be used to more efficiently find whether two sets intersect.

This takes a 1000 piece 100% strictness least merges order puzzle's generate_early from 10s to 7.5s for me, and produces identical output with the same seed.

### [Track the number of added pieces instead of calculating it each time](https://github.com/spineraks-org/ArchipelagoJigsaw/commit/b1e9dd29400bc92b942eca5bfaac41053add5f5e)

The number of pieces in the current clusters can be tracked and incremented every time a piece is added.

This avoids recalculating the more costly operation of counting the total number of pieces in the current clusters each time.

This takes a 1000 piece 100% strictness least merges order puzzle's generate_early from 7.5s to 5s for me, and produces identical output with the same seed.

The unused `remove_piece()` function has been modified, but has not been tested.

### [Skip creating a new solution when only the number of merges is needed](https://github.com/spineraks-org/ArchipelagoJigsaw/pull/2/commits/1948ed03999967b83fc542e6978ad333de72314f)

This removes some of the work that add_piece was doing that is not necessary for testing if a piece is suitable to create a new solution given a specific piece_order.

This change alone reduces takes a 1000 piece 100% strictness least merges order puzzle's generate_early from 5s to 4.4s for me, and produces identical output with the same seed.

However, the iteration and summation can now be performed in optimised C code using `sum` and `map`, further reducing from 4.4s to 2.8s.

## How was this tested?

Generations with a fixed seed with a 1000 piece random order yaml, a 1000 piece least merges order yaml, and a 1000 piece every piece fits order yaml produced the same result before and after, but with better `generate_early()` performance.