@Mysteryem 2025-04-02
An alternative to [Track the number of added pieces instead of calculating it each time](https://github.com/spineraks-org/ArchipelagoJigsaw/commit/b1e9dd29400bc92b942eca5bfaac41053add5f5e) would be to change `add_piece` to return the number of new merges rather than the total number of merges. Performance-wise, this is effectively identical and produces the same results. The main downside is that it would make the unused `remove_piece()` function slower, since it would have to calculate the number of merges.

This alternative, applied on top of the changes in this PR, can be found at https://github.com/Mysteryem/Archipelago-ahit/commit/adda303bdf19733bd3aba4b68404ac21a24680f6.