@Mysteryem 2025-04-02
Generating with 3 1000 piece puzzles with each piece_order at 100% strictness and random piece_order_type.

An example before:
```
Took 0.5222 seconds in JigsawWorld.generate_early for player 1, named EveryPieceFits.
Took 14.6960 seconds in JigsawWorld.generate_early for player 2, named LeastMerges.
Took 0.0750 seconds in JigsawWorld.generate_early for player 3, named RandomOrder.
```

An example after:
```
Took 0.1114 seconds in JigsawWorld.generate_early for player 1, named EveryPieceFits.
Took 4.8739 seconds in JigsawWorld.generate_early for player 2, named LeastMerges.
Took 0.0405 seconds in JigsawWorld.generate_early for player 3, named RandomOrder.
```