@Mysteryem 2025-04-02
Updated gen example, on the same seed as before, generating with 3 1000 piece puzzles with each piece_order at 100% strictness and `piece_order_type: random_order`.
```
Took 0.1143 seconds in JigsawWorld.generate_early for player 1, named EveryPieceFits.
Took 2.7945 seconds in JigsawWorld.generate_early for player 2, named LeastMerges.
Took 0.0414 seconds in JigsawWorld.generate_early for player 3, named RandomOrder.
```

This shows a small increase for `EveryPieceFits` and `RandomOrder`, but this is just noise that would be eliminated by performing many generations and taking an average.