@Mysteryem 2025-04-05
Superseded by #3 due to its better performance.